requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
PyJWT
sqlparse
psycopg2-binary
python-dotenv
django-cors-headers
Pillow